final review

+++ rmall.c
@@ -20,20 +20,14 @@ int is_file(const char *path){
 }
 
 int remove_directory(const char *fpath, const struct stat *sb, int typeflag, struct FTW *flag){
-    if(is_file(fpath))
-        unlink(fpath);
-    else if(is_directory(fpath))
-        rmdir(fpath);
+    if(is_file(fpath)) unlink(fpath);
+    else if(is_directory(fpath)) rmdir(fpath);
     return 0; //no error
 }
 
 int main(int argc, char* argv[]){
-    if(argc != 2)
-        fprintf(stdout, "<rmall> prend un chemin relatif/absolu en parametre\n");
-    else
-    {
-        if (!argv[1] || access(argv[1], R_OK)) perror("ce parametre n'est pas accessbile");
-        else if (nftw(argv[1], &remove_directory, 1, FTW_DEPTH)) perror("error occurred in ftw");
-    }
+    if(argc != 2) fprintf(stdout, "<rmall> prend un chemin relatif/absolu en parametre\n");
+    else if (!argv[1] || access(argv[1], R_OK)) perror("ce parametre n'est pas accessbile");
+    else if (nftw(argv[1], &remove_directory, 1, FTW_DEPTH)) perror("error occurred in ftw");
     return EXIT_SUCCESS;
 }
